backend/requirements.txt
fastapi
uvicorn
librosa
tensorflow-cpu
numpy